requirements.txt
pyyaml
tqdm
transformers>=4.30.0
numpy<2.0.0
google-generativeai
anthropic
openai
accelerate
sentencepiece
argparse
